Refactor Makefile and dunst configuration
This commit updates the Makefile to use more consistent variable names
and adds an 'icons' target to copy dunst icons to the user's data
directory. It also modifies the dunst configuration file to change
the background and foreground colors, set a default icon, and modify
the icon path to include custom icons. Finally, it changes the keybindings
for volume control to use a custom script and adds a bind command for
toggling mute.