Add brightness stack to dunstrc
This commit adds a new stack to dunstrc named "brightness" to track
notifications related to changing screen brightness. The stack is
configured to timeout after 2 seconds.

It also updates sxhkdrc to use a custom script located at
$HOME/.local/bin/change-brightness for changing the screen brightness.
The script accepts two arguments, the first being the percentage by which
to change the brightness and the second being either "+" or "-" for increasing
or decreasing the brightness respectively. This enables more precise control
over screen brightness compared to using the 'brightnessctl' command.