feat: Add shellstart function to bashrc and zshrc
This commit adds a new function called `shellstart` to the `bashrc` and `zshrc` files. The `shellstart` function checks for the availability of either `neofetch` or `fastfetch` commands and executes the first one found. This function is added to enhance the shell startup experience by displaying system information using a fetch tool.

Additionally, a new image file `Arch-linux-logo.png` is added to the `neofetch` directory.

The changes made in this commit are as follows:

- In `bash/bashrc`:
  - Added the `shellstart` function.
  - Removed the conditional block that checks for the availability of `fastfetch` and `neofetch` commands.

- In `neofetch/config.conf`:
  - Modified the `image_backend` setting to use the `kitty` backend for displaying images.
  - Modified the `image_source` setting to point to the `Arch-linux-logo.png` image file.

- In `zsh/.zshrc`:
  - Added the `neofetch` command to display system information at shell startup.

Note: The `neofetch` command is assumed to be available in the system. If it is not present, the `fastfetch` command will be used instead.